feat: verify jwt
